+++ R/chk.R
@@ -23,6 +23,6 @@ chk_term <- function(x, validate = "complete", x_name = NULL) {
   if (!all(valid_term(x))) {
     abort_chk("All elements of term vector ", x_name, " must be valid")
   }
-  if (is.inconsistent_terms(x)) abort_chk("All elements of term vector ", x_name, " must be consistent")
+  if (is_inconsistent_terms(x)) abort_chk("All elements of term vector ", x_name, " must be consistent")
   abort_chk("All elements of term vector ", x_name, " must be complete")
 }