+++ R/pdims.R
@@ -18,7 +18,7 @@ universals::pdims
 #' pdims(as.term(c("alpha[1]", "alpha[3]", "beta[1,1]", "beta[2,1]")))
 pdims.term <- function(x, ...) {
   chk_not_any_na(x)
-  if (is.inconsistent_terms(x)) {
+  if (is_inconsistent_terms(x)) {
     err("`x` must have terms with consistent parameter dimensions.")
   }
   x <- sort(x)